+++ src/js/TwbsToggleButtons.js
@@ -1,4 +1,4 @@
-/** @preserve Twitter Bootstrap Toogle Buttons 0.0.2
+/** @preserve Twitter Bootstrap Toogle Buttons 0.0.3
  * Available under the MIT license.
  * See https://github.com/prokki/twbs-toggle-buttons for more information.
  */
@@ -260,6 +260,18 @@ class TwbsToggleButtons
 
 		$(button).find(":input").prop("checked", true);
 		$(button).find(":input").attr("checked", "checked");
+
+		// workaround on radio and checkbox button:
+		// the attribute "aria-pressed" stays on "false" (even if the attribute is changed furthermore),
+		// add the attribute manually with setTimeout() function 
+		if ((this._getInputType() === TwbsToggleButtons.TYPE_RADIO() || this._getInputType() === TwbsToggleButtons.TYPE_CHECKBOX()) && button.getAttribute("aria-pressed") === "false" )
+		{
+			window.setTimeout(function ()
+			{
+				button.classList.remove(TwbsToggleButtons.ACTIVE_CLASS());
+				button.setAttribute("aria-pressed", "true");
+			}, 0);
+		}
 	};
 
 	/**
@@ -287,10 +299,10 @@ class TwbsToggleButtons
 		$(button).find(":input").prop("checked", false);
 		$(button).find(":input").attr("checked", null);
 
-		// workaround on radio button:
-		//   the attribute "aria-pressed" stays on "true" (even if the attribute is changed furthermore),
-		//   remove the attribute manually with setTimeout() function 
-		if ( this._getInputType() === TwbsToggleButtons.TYPE_RADIO() && button.getAttribute("aria-pressed") === "true" )
+		// workaround on radio and checkbox button:
+		// the attribute "aria-pressed" stays on "true" (even if the attribute is changed furthermore),
+		// remove the attribute manually with setTimeout() function 
+		if ((this._getInputType() === TwbsToggleButtons.TYPE_RADIO() || this._getInputType() === TwbsToggleButtons.TYPE_CHECKBOX()) && button.getAttribute("aria-pressed") === "true" )
 		{
 			window.setTimeout(function ()
 			{