+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### [0.0.3] - 2019-07-29
+
+### Fixed
+
+* Workaround for checkbox buttons when using latest bootstrap (not alpha): Remove attribute `"aria-pressed"` and class `active` manually to deactivate a checkbox button.
+* Workaround for radio and checkbox buttons when using latest bootstrap (not alpha): Add attribute `"aria-pressed"` and class `active` manually to activate a radio and checkbox button.
+
 ### [0.0.2] - 2017-05-11
 
 #### Added